Introduced AnemoiBaseModel class
